+++ core/launcher/api_container_launcher/api_container_launcher.go
@@ -5,19 +5,17 @@
 package api_container_launcher
 
 import (
-	"bytes"
 	"context"
 	"fmt"
-	"net"
-	"time"
-
 	"github.com/google/martian/log"
-	"github.com/kurtosis-tech/container-engine-lib/lib/backend_impls/docker/docker_manager"
+	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface"
+	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/api_container"
+	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/enclave"
+	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/port_spec"
 	"github.com/kurtosis-tech/kurtosis-core/launcher/args"
-	"github.com/kurtosis-tech/kurtosis-core/launcher/enclave_container_launcher"
-	"github.com/kurtosis-tech/object-attributes-schema-lib/schema"
 	"github.com/kurtosis-tech/stacktrace"
 	"github.com/sirupsen/logrus"
+	"net"
 )
 
 const (
@@ -25,48 +23,26 @@ const (
 	DefaultVersion = "1.41.2"
 	// !!!!!!!!!!!!!!!!!! DO NOT MODIFY THIS! IT WILL BE UPDATED AUTOMATICALLY DURING THE RELEASE PROCESS !!!!!!!!!!!!!!!
 
-	portProtocolToMonitorWhenWaitingForAvailability = "tcp"
-
-	maxWaitForAvailabilityRetries         = 10
-	timeBetweenWaitForAvailabilityRetries = 1 * time.Second
-
-	availabilityWaitingExecCmdSuccessExitCode = 0
-
 	// The location where the enclave data directory (on the Docker host machine) will be bind-mounted
 	//  on the API container
 	enclaveDataDirpathOnAPIContainer = "/kurtosis-enclave-data"
 
 	// TODO This should come from the same logic that builds the server image!!!!!
 	containerImage = "kurtosistech/kurtosis-core_api"
-
-	// We always want the user to be running the latest version of the specified API container image
-	shouldPullImageBeforeLaunching = true
-
-	// The API container must have the Docker socket bind-mounted so that it can start other containers
-	shouldBindMountDockerSocket = true
-
-	// API container doesn't have an alias
-	containerAlias = ""
 )
 
-// All the following are set to the value of "don't use these"
-var entrypointArgs []string = nil
-var cmdArgs []string = nil
-var volumeMounts map[string]string = nil
-
 type ApiContainerLauncher struct {
-	enclaveContainerLauncher *enclave_container_launcher.EnclaveContainerLauncher
-	dockerManager            *docker_manager.DockerManager
+	kurtosisBackend backend_interface.KurtosisBackend
 }
 
-func NewApiContainerLauncher(enclaveContainerLauncher *enclave_container_launcher.EnclaveContainerLauncher, dockerManager *docker_manager.DockerManager) *ApiContainerLauncher {
-	return &ApiContainerLauncher{enclaveContainerLauncher: enclaveContainerLauncher, dockerManager: dockerManager}
+func NewApiContainerLauncher(kurtosisBackend backend_interface.KurtosisBackend) *ApiContainerLauncher {
+	return &ApiContainerLauncher{kurtosisBackend: kurtosisBackend}
 }
 
 func (launcher ApiContainerLauncher) LaunchWithDefaultVersion(
 	ctx context.Context,
 	logLevel logrus.Level,
-	enclaveId string,
+	enclaveId enclave.EnclaveID,
 	networkId string,
 	subnetMask string,
 	grpcListenPort uint16,
@@ -78,13 +54,10 @@ func (launcher ApiContainerLauncher) LaunchWithDefaultVersion(
 	metricsUserID string,
 	didUserAcceptSendingMetrics bool,
 ) (
-	resultContainerId string,
-	resultPublicIpAddr net.IP,
-	resultGrpcPublicPort *enclave_container_launcher.EnclaveContainerPort,
-	resultGrpcProxyPublicPort *enclave_container_launcher.EnclaveContainerPort,
+	resultApiContainer *api_container.APIContainer,
 	resultErr error,
 ) {
-	containerId, publicIpAddr, publicGrpcPort, publicGrpcProxyPort, err := launcher.LaunchWithCustomVersion(
+	resultApiContainer, err := launcher.LaunchWithCustomVersion(
 		ctx,
 		DefaultVersion,
 		logLevel,
@@ -101,16 +74,16 @@ func (launcher ApiContainerLauncher) LaunchWithDefaultVersion(
 		didUserAcceptSendingMetrics,
 	)
 	if err != nil {
-		return "", nil, nil, nil, stacktrace.Propagate(err, "An error occurred launching the API container with default version tag '%v'", DefaultVersion)
+		return nil, stacktrace.Propagate(err, "An error occurred launching the API container with default version tag '%v'", DefaultVersion)
 	}
-	return containerId, publicIpAddr, publicGrpcPort, publicGrpcProxyPort, nil
+	return resultApiContainer, nil
 }
 
 func (launcher ApiContainerLauncher) LaunchWithCustomVersion(
 	ctx context.Context,
 	imageVersionTag string,
 	logLevel logrus.Level,
-	enclaveId string,
+	enclaveId enclave.EnclaveID,
 	networkId string,
 	subnetMask string,
 	grpcPortNum uint16,
@@ -122,28 +95,9 @@ func (launcher ApiContainerLauncher) LaunchWithCustomVersion(
 	metricsUserID string,
 	didUserAcceptSendingMetrics bool,
 ) (
-	resultContainerId string,
-	resultPublicIpAddr net.IP,
-	resultGrpcPublicPort *enclave_container_launcher.EnclaveContainerPort,
-	resultGrpcProxyPublicPort *enclave_container_launcher.EnclaveContainerPort,
+	resultApiContainer *api_container.APIContainer,
 	resultErr error,
 ) {
-	objAttrsSupplier := func(enclaveObjAttrsProvider schema.EnclaveObjectAttributesProvider) (schema.ObjectAttributes, error) {
-		apiContainerAttrs, err := enclaveObjAttrsProvider.ForApiContainer(
-			apiContainerIpAddr,
-			grpcPortNum,
-			grpcProxyPortNum,
-		)
-		if err != nil {
-			return nil, stacktrace.Propagate(
-				err,
-				"An error occurred getting the API container object attributes using port num '%v' and proxy port '%v'",
-				grpcPortNum,
-				grpcProxyPortNum,
-			)
-		}
-		return apiContainerAttrs, nil
-	}
 
 	takenIpAddrStrSet := map[string]bool{
 		gatewayIpAddr.String():      true,
@@ -154,7 +108,7 @@ func (launcher ApiContainerLauncher) LaunchWithCustomVersion(
 		logLevel.String(),
 		grpcPortNum,
 		grpcProxyPortNum,
-		enclaveId,
+		string(enclaveId),
 		networkId,
 		subnetMask,
 		apiContainerIpAddr.String(),
@@ -166,12 +120,12 @@ func (launcher ApiContainerLauncher) LaunchWithCustomVersion(
 		didUserAcceptSendingMetrics,
 	)
 	if err != nil {
-		return "", nil, nil, nil, stacktrace.Propagate(err, "An error occurred creating the API container args")
+		return nil, stacktrace.Propagate(err, "An error occurred creating the API container args")
 	}
 
 	envVars, err := args.GetEnvFromArgs(argsObj)
 	if err != nil {
-		return "", nil, nil, nil, stacktrace.Propagate(err, "An error occurred generating the API container's environment variables")
+		return nil, stacktrace.Propagate(err, "An error occurred generating the API container's environment variables")
 	}
 
 	containerImageAndTag := fmt.Sprintf(
@@ -180,117 +134,30 @@ func (launcher ApiContainerLauncher) LaunchWithCustomVersion(
 		imageVersionTag,
 	)
 
-	grpcPort, err := enclave_container_launcher.NewEnclaveContainerPort(grpcPortNum, enclave_container_launcher.EnclaveContainerPortProtocol_TCP)
+	// Create Enclave API container
+	grpcPort, err := port_spec.NewPortSpec(grpcPortNum, port_spec.PortProtocol_TCP)
 	if err != nil {
-		return "", nil, nil, nil, stacktrace.Propagate(err, "An error occurred constructing the enclave container port object representing the API container's gRPC port '%v'", grpcPortNum)
+		stacktrace.Propagate(err, "An error occurred creating a port spec for GRPC port '%v'", grpcPortNum)
 	}
-
-	grpcProxyPort, err := enclave_container_launcher.NewEnclaveContainerPort(grpcProxyPortNum, enclave_container_launcher.EnclaveContainerPortProtocol_TCP)
+	grpcProxyPort, err := port_spec.NewPortSpec(grpcProxyPortNum, port_spec.PortProtocol_TCP)
 	if err != nil {
-		return "", nil, nil, nil, stacktrace.Propagate(err, "An error occurred constructing the enclave container port object representing the API container's gRPC port with portNum '%v' and grpcPortNum '%v'", grpcPortNum, grpcProxyPortNum)
-	}
-
-	privatePorts := map[string]*enclave_container_launcher.EnclaveContainerPort{
-		schema.KurtosisInternalContainerGRPCPortID:      grpcPort,
-		schema.KurtosisInternalContainerGRPCProxyPortID: grpcProxyPort,
+		stacktrace.Propagate(err, "An error occurred creating a port spec for GRPC Proxy port '%v'", grpcProxyPortNum)
 	}
 
 	log.Debugf("Launching Kurtosis API container...")
-	containerId, publicIpAddr, publicPorts, err := launcher.enclaveContainerLauncher.Launch(
+	apiContainer, err := launcher.kurtosisBackend.CreateAPIContainer(
 		ctx,
 		containerImageAndTag,
-		shouldPullImageBeforeLaunching,
+		enclaveId,
 		apiContainerIpAddr,
-		networkId,
-		enclaveDataDirpathOnAPIContainer,
-		privatePorts,
-		objAttrsSupplier,
+		grpcPort.GetNumber(),
+		grpcProxyPort.GetNumber(),
+		enclaveDataDirpathOnHostMachine,
 		envVars,
-		shouldBindMountDockerSocket,
-		containerAlias,
-		entrypointArgs,
-		cmdArgs,
-		volumeMounts,
 	)
 	if err != nil {
-		return "", nil, nil, nil, stacktrace.Propagate(err, "An error occurred launching the API container")
+		return nil, stacktrace.Propagate(err, "An error occurred launching the API container")
 	}
 
-	shouldKillContainer := true
-	defer func() {
-		if shouldKillContainer {
-			if killErr := launcher.dockerManager.KillContainer(context.Background(), containerId); killErr != nil {
-				logrus.Errorf("The function to create the API container didn't finish successful so we tried to kill the container we created, but the killing threw an error:")
-				logrus.Error(killErr)
-			}
-		}
-	}()
-
-	if err := waitForAvailability(ctx, launcher.dockerManager, containerId, grpcPortNum); err != nil {
-		return "", nil, nil, nil, stacktrace.Propagate(err, "An error occurred waiting for the API container's grpc port to become available")
-	}
-
-	if err := waitForAvailability(ctx, launcher.dockerManager, containerId, grpcProxyPortNum); err != nil {
-		return "", nil, nil, nil, stacktrace.Propagate(err, "An error occurred waiting for the API container's grpc-proxy port to become available")
-	}
-
-	publicGrpcPort, found := publicPorts[schema.KurtosisInternalContainerGRPCPortID]
-	if !found {
-		return "", nil, nil, nil, stacktrace.NewError("No public port was found for '%v' - this is very strange!", schema.KurtosisInternalContainerGRPCPortID)
-	}
-
-	publicGrpcProxyPort, found := publicPorts[schema.KurtosisInternalContainerGRPCProxyPortID]
-	if !found {
-		return "", nil, nil, nil, stacktrace.NewError("No public port was found for '%v' - this is very strange!", schema.KurtosisInternalContainerGRPCProxyPortID)
-	}
-
-	shouldKillContainer = false
-	return containerId, publicIpAddr, publicGrpcPort, publicGrpcProxyPort, nil
-}
-
-func waitForAvailability(ctx context.Context, dockerManager *docker_manager.DockerManager, containerId string, listenPortNum uint16) error {
-	commandStr := fmt.Sprintf(
-		"[ -n \"$(netstat -anp %v | grep LISTEN | grep %v)\" ]",
-		portProtocolToMonitorWhenWaitingForAvailability,
-		listenPortNum,
-	)
-	execCmd := []string{
-		"sh",
-		"-c",
-		commandStr,
-	}
-	for i := 0; i < maxWaitForAvailabilityRetries; i++ {
-		outputBuffer := &bytes.Buffer{}
-		exitCode, err := dockerManager.RunExecCommand(ctx, containerId, execCmd, outputBuffer)
-		if err == nil {
-			if exitCode == availabilityWaitingExecCmdSuccessExitCode {
-				return nil
-			}
-			logrus.Debugf(
-				"API container availability-waiting command '%v' returned without a Docker error, but exited with non-%v exit code '%v' and logs:\n%v",
-				commandStr,
-				availabilityWaitingExecCmdSuccessExitCode,
-				exitCode,
-				outputBuffer.String(),
-			)
-		} else {
-			logrus.Debugf(
-				"API container availability-waiting command '%v' experienced a Docker error:\n%v",
-				commandStr,
-				err,
-			)
-		}
-
-		// Tiny optimization to not sleep if we're not going to run the loop again
-		if i < maxWaitForAvailabilityRetries {
-			time.Sleep(timeBetweenWaitForAvailabilityRetries)
-		}
-	}
-
-	return stacktrace.NewError(
-		"The API container didn't become available (as measured by the command '%v') even after retrying %v times with %v between retries",
-		commandStr,
-		maxWaitForAvailabilityRetries,
-		timeBetweenWaitForAvailabilityRetries,
-	)
+	return apiContainer, nil
 }